minor changes from PR comments
